Retrieve the friend list of a user.
It is now possible to retrieve the friendlist of a user and add users to an account.  This code can run without problems except the known issue mentioned below. 

ISSUE: Friendlist currently only adds itself, so it will need to be redirected to hold information from another user. 

What's Next: 
1. -Resolve the issue to add the correct Friend UUID to the list of friends for the target user. 
2. -Input the remove friend issue.  (This should not be that challenging.  List.remove(). (1) seems like the more challenging task.  )
3. -Clean up code.